refactor: Modify post method repository
