+++ src/services/post.service.ts
@@ -1,0 +1,33 @@
+import { Post } from "../types/post.interface";
+
+export const getLocationPosts = async (areaId: number): Promise<Post[]> => {
+  const response = await fetch("http://localhost:3004/post");
+  if (!response.ok) {
+    throw new Error("エラーが発生しました。");
+  }
+
+  const data: Post[] = await response.json();
+
+  if (data) {
+    const matchedPosts = data.filter(
+      (post: Post) => Number(post.postAreaId) === areaId
+    );
+    return matchedPosts;
+  } else {
+    throw new Error("エラーが発生しました。");
+  }
+};
+
+export const getUserPosts = async (userId: number) => {
+  const res = await fetch(`http://localhost:3004/post`);
+
+  if (!res.ok) {
+    throw new Error("エラーが発生しました。");
+  }
+
+  const posts = await res.json();
+
+  const userPosts = posts.filter((post: Post) => post.userId === userId);
+
+  return userPosts;
+};
