+++ src/services/user.service.ts
@@ -1,5 +1,4 @@
 import { User, UserUpdate } from "../types/user.interface";
-import { Post } from "../types/post.interface";
 
 export const getUser = async (id: string): Promise<User> => {
   const res = await fetch(`http://localhost:3004/user/${id}`);
@@ -45,18 +44,4 @@ export const uploadProfileImage = async (file: File): Promise<string> => {
   const data = await res.json();
 
   return data.imageUrl;
-};
-
-export const getUserPost = async (userId: number) => {
-  const res = await fetch(`http://localhost:3004/post`);
-
-  if (!res.ok) {
-    throw new Error("エラーが発生しました。");
-  }
-
-  const posts = await res.json();
-
-  const userPosts = posts.filter((post: Post) => post.userId === userId);
-
-  return userPosts;
 };
